requirements.txt
# Money-Game Lending Platform - Python Dependencies

# Supabase Client
supabase==2.3.0

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Date/Time Utilities
python-dateutil==2.8.2

# Security & SSL Verification
certifi==2023.11.17
urllib3==2.1.0

# LINE Messaging API (optional, for advanced features)
line-bot-sdk==3.5.0
